+++ src/games/abstracts/components/ADisplayableComponent.cpp
@@ -8,7 +8,7 @@
 #include "ADisplayableComponent.hpp"
 
 ADisplayableComponent::ADisplayableComponent(
-    shared::types::Vector2i position,
+    shared::types::Vector2f position,
             shared::types::Vector2u size, entity::IEntity &entity, unsigned int zIndex,
             components::ComponentType type = components::ComponentType::DISPLAYABLE)
     : PositionableComponent(position, size, entity, type),
