+++ src/games/abstracts/components/ADisplayableComponent.hpp
@@ -13,7 +13,7 @@
 class ADisplayableComponent : public virtual components::IDisplayableComponent,
                               public PositionableComponent {
     public:
-        ADisplayableComponent(shared::types::Vector2i position,
+        ADisplayableComponent(shared::types::Vector2f position,
             shared::types::Vector2u size, entity::IEntity &entity, unsigned int zIndex,
             components::ComponentType type);
         ~ADisplayableComponent();
