+++ src/games/solarfox/entities/enemy/SolarFoxEnemy.cpp
@@ -7,7 +7,7 @@
 
 #include "SolarFoxEnemy.hpp"
 
-SolarFoxEnemy::SolarFoxEnemy(shared::types::Vector2i position, shared::types::Vector2u size,
+SolarFoxEnemy::SolarFoxEnemy(shared::types::Vector2f position, shared::types::Vector2u size,
     shared::types::Vector2u origin)
 {
     std::shared_ptr<TextureComponent> texture = std::make_shared<TextureComponent>(
