+++ src/graphicals/sdl2/Window/SDL2Window.cpp
@@ -321,14 +321,14 @@ std::vector<std::shared_ptr<shared::graphics::events::IEvent>> SDL2Window::getEv
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonPressEvent>
                     (shared::graphics::events::IMouseButtonEvent::RIGHT,
-                    (Vector2i) {(event.button.x / static_cast<int>(sizeTiles.x)),
-                    (event.button.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.button.x / (sizeTiles.x)),
+                    (event.button.y / (sizeTiles.y))}));
                 } else if (event.button.button == SDL_BUTTON_LEFT) {
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonPressEvent>
                     (shared::graphics::events::IMouseButtonEvent::LEFT,
-                    (Vector2i) {(event.button.x / static_cast<int>(sizeTiles.x)),
-                    (event.button.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.button.x / (sizeTiles.x)),
+                    (event.button.y / (sizeTiles.y))}));
                 }
                 break;
             case SDL_MOUSEBUTTONUP:
@@ -336,20 +336,20 @@ std::vector<std::shared_ptr<shared::graphics::events::IEvent>> SDL2Window::getEv
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonReleaseEvent>
                     (shared::graphics::events::IMouseButtonEvent::RIGHT,
-                    (Vector2i) {(event.button.x / static_cast<int>(sizeTiles.x)),
-                    (event.button.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.button.x / (sizeTiles.x)),
+                    (event.button.y / (sizeTiles.y))}));
                 } else if (event.button.button == SDL_BUTTON_LEFT) {
                     eventsList.push_back
                     (std::make_shared<shared::graphics::events::MouseButtonReleaseEvent>
                     (shared::graphics::events::IMouseButtonEvent::LEFT,
-                    (Vector2i) {(event.button.x / static_cast<int>(sizeTiles.x)),
-                    (event.button.y / static_cast<int>(sizeTiles.y))}));
+                    (Vector2f) {(event.button.x / (sizeTiles.x)),
+                    (event.button.y / (sizeTiles.y))}));
                 }
                 break;
             case SDL_MOUSEMOTION:
                 eventsList.push_back(std::make_shared<shared::graphics::events::MouseMoveEvent>
-                    ((Vector2i) {(event.motion.x / static_cast<int>(sizeTiles.x)),
-                    (event.motion.y / static_cast<int>(sizeTiles.y))}));
+                    ((Vector2f) {(event.motion.x / (sizeTiles.x)),
+                    (event.motion.y / (sizeTiles.y))}));
                 break;
             case SDL_KEYDOWN:
                 keycode = event.key.keysym.sym;
